+++ main.rb
@@ -1,10 +1,11 @@
 require './app'
-
+require "./select_option"
 def main
   app = App.new
+  select_obj = SelectOption.new()
   loop do
     show_menu
-    select_option(app)
+    select_obj.select_option(app)
   end
 end
 
@@ -19,28 +20,4 @@ def show_menu
   puts '7. Quit'
 end
 
-def select_option(app)
-  options_hash = {
-    '1' => :books_list,
-    '2' => :people_list,
-    '3' => :create_book,
-    '4' => :list_person_id_of_rentals,
-    '5' => :create_person,
-    '6' => :create_rental,
-    '7' => :exit_application
-  }
-
-  input = gets.chomp
-  if options_hash.key?(input)
-    app.send(options_hash[input])
-  else
-    puts 'INVALID!! Please select again!'
-  end
-end
-
-def exit_application
-  puts 'Goodbye! see you again....'
-  exit
-end
-
-main
+main()
